requirements.txt
chromadb
openai
langchain
python-dotenv
tiktoken
wikipedia
pypdf
docx2txt
streamlit